requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
aiohttp==3.9.3
pydantic==2.5.2
gunicorn==21.2.0 